+++ CHANGES.rst
@@ -33,6 +33,7 @@ Development
 - Make ``make setup`` more platform-agnostic, resolve #20
 - Add external antenna option for vanilla MicroPython
 - Use trimmed-down settings.py for PYBD
+- Add configuration settings overlay per ``settings-user.json``
 
 
 2019-08-19 0.6.0